requirements.txt
pyrogram
tgcrypto
aiohttp
python-dotenv
motor
aiofiles
shortzy
dnspython
apscheduler
requests
psutil
pytz
pyromod
requests
cloudscraper
bs4 
lxml
cfscrape
urllib3
flask
gunicorn
